[Bugfix] Fix duplicate SHM cache initialization
